Minor NIF fixes and cleanup

See merge request OpenMW/openmw!377

+++ components/nif/niffile.cpp
@@ -160,7 +160,7 @@ void NIFFile::parse(Files::IStreamPtr stream)
         userVer = nif.getUInt();
 
     // Number of records
-    size_t recNum = nif.getUInt();
+    unsigned int recNum = nif.getUInt();
     records.resize(recNum);
 
     // Bethesda stream header
@@ -212,7 +212,7 @@ void NIFFile::parse(Files::IStreamPtr stream)
     }
 
     const bool hasRecordSeparators = ver >= NIFStream::generateVersion(10,0,0,0) && ver < NIFStream::generateVersion(10,2,0,0);
-    for(size_t i = 0;i < recNum;i++)
+    for (unsigned int i = 0; i < recNum; i++)
     {
         Record *r = nullptr;
 
@@ -253,11 +253,11 @@ void NIFFile::parse(Files::IStreamPtr stream)
         r->read(&nif);
     }
 
-    size_t rootNum = nif.getUInt();
+    unsigned int rootNum = nif.getUInt();
     roots.resize(rootNum);
 
     //Determine which records are roots
-    for(size_t i = 0;i < rootNum;i++)
+    for (unsigned int i = 0; i < rootNum; i++)
     {
         int idx = nif.getInt();
         if (idx >= 0 && idx < int(records.size()))
